+++ src/main/java/CRUD/dao/UserDao.java
@@ -7,9 +7,9 @@ import java.util.List;
 public interface UserDao {
     void addUser(User user);
 
-    void removeUser(int id);
+    void removeUser(Long id);
 
-    User getUserById(int id);
+    User getUserById(Long id);
 
     List<User> userList();
