+++ src/main/java/CRUD/models/User.java
@@ -13,7 +13,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private Long id;
 
     @Column(name = "name")
     @NotEmpty(message = "Name should not be empty")
@@ -42,7 +42,7 @@ public class User {
     public User() {
     }
 
-    public User(int id, String name, String surname, int age, String email, String city) {
+    public User(Long id, String name, String surname, int age, String email, String city) {
         this.id = id;
         this.name = name;
         this.surname = surname;
@@ -51,11 +51,11 @@ public class User {
         this.city = city;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
